Move `Attributes` tests to core-common
Move `Attributes` tests from sdk-common to core-common. To achieve
this, `sdk-common` now depends on `core-common % "test->test"`.
