+++ core/common/src/test/scala/org/typelevel/otel4s/AttributesLawTests.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Typelevel
+ * Copyright 2022 Typelevel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,11 @@
  */
 
 package org.typelevel.otel4s
-package sdk
 
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.MonoidTests
 import munit.DisciplineSuite
-import org.typelevel.otel4s.sdk.arbitrary._
+import org.typelevel.otel4s.arbitrary._
 
 class AttributesLawTests extends DisciplineSuite {
