+++ core/common/src/test/scala/org/typelevel/otel4s/AttributesProps.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Typelevel
+ * Copyright 2022 Typelevel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
  */
 
 package org.typelevel.otel4s
-package sdk
 
 import cats.Show
 import cats.syntax.semigroup._
@@ -23,8 +22,8 @@ import munit.ScalaCheckSuite
 import org.scalacheck.Arbitrary
 import org.scalacheck.Gen
 import org.scalacheck.Prop.forAll
-import org.typelevel.otel4s.sdk.arbitrary.attribute
-import org.typelevel.otel4s.sdk.arbitrary.attributes
+import org.typelevel.otel4s.arbitrary.attribute
+import org.typelevel.otel4s.arbitrary.attributes
 
 class AttributesProps extends ScalaCheckSuite {
